Moved SetupCredentials out of util.go and into vcs_secret.go
